requirements.txt
aiohttp
allure-robotframework
awscli
axe-selenium-python
beautifulsoup4
boto3
botocore
browserstack-sdk
configobj
coverage
cryptography
docx
faiss-cpu
fitz
frontend
imagehash
langchain
langchain-huggingface
language_tool_python
lighthouse
lxml
matplotlib
mock
nest_asyncio
nltk
numpy
openai
opencv-python
openpyxl
oracledb
pandas
pillow
pillow-avif-plugin
plotly
prometheus-client
psutil
PyJWT
pymupdf
pyotp
PyPDF2
pyppeteer
pytesseract
pytest-axe
python-docx
python-pptx
reportlab
requests
robotframework
robotframework-dashboard
robotframework-databaselibrary
robotframework-dependencylibrary
robotframework-extendedselenium2library
robotframework-heal
robotframework-jsonlibrary
robotframework-lint
robotframework-mcp
robotframework-metrics
robotframework-pabot
robotframework-requests
robotframework-robocop
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary
robotframework-selfhealing-agents
robotframework-tidy
scikit-image
seaborn
selenium
spacy
streamlit
testfixtures
torch
urllib3
websockets
yarl